% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_intervals.r
\name{gsvb.credible_intervals}
\alias{gsvb.credible_intervals}
\title{Construct the marginal credible intervals for the model coefficients}
\usage{
gsvb.credible_intervals(fit, prob = 0.95)
}
\arguments{
\item{fit}{fit model}

\item{prob}{the probability contained in the credible interval of each coefficient}
}
\value{
matrix with the columns for:
\item{lower}{the lower bound of the credible interval}
\item{upper}{the upper bound of the credible interval}
\item{contains.dirac}{indicates whether the Dirac mass at zero is contained in the interval}
}
\description{
Construct the marginal credible intervals for the model coefficients
}
\section{Details}{
 
Returns the interval of highest posterior density containing prob * 100% of the mass
}

\examples{
n <- 100
p <- 1000
gsize <- 5
groups <- c(rep(1:(p/gsize), each=gsize))

X <- matrix(rnorm(n * p), nrow=n, ncol=p)
b <- c(rep(0, gsize), rep(-4, gsize), rep(8, gsize), rep(0, p - 3 * gsize))
y <- X \%*\% b + rnorm(n, 0, 1)

f <- gsvb.fit(y, X, groups)
credible.intervals(f, prob=0.95)

}
