% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{gsvb.fit}
\alias{gsvb.fit}
\title{Fit high-dimensional group-sparse regression models}
\usage{
gsvb.fit(
  y,
  X,
  groups,
  intercept = TRUE,
  diag_covariance = TRUE,
  lambda = 1,
  a0 = 1,
  b0 = length(unique(groups)),
  tau_a0 = 0.001,
  tau_b0 = 0.001,
  mu = NULL,
  s = apply(X, 2, function(x) 1/sqrt(sum(x^2) * tau_a0/tau_b0 + 2 * lambda)),
  g = rep(0.5, ncol(X)),
  track_elbo = TRUE,
  track_elbo_every = 5,
  track_elbo_mcn = 500,
  niter = 150,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{y}{response vector.}

\item{X}{input matrix.}

\item{groups}{group structure.}

\item{intercept}{should an intercept term be included.}

\item{diag_covariance}{should the covariance matrix used in the variational approximation be diagonal. Note: if true then the *standard deviations* for each coefficient are returned. If false then covariance matrices for each group are returned.}

\item{lambda}{penalization hyperparameter for the multivariate exponential prior.}

\item{a0}{shape parameter for the Beta(a0, b0) mixing prior.}

\item{b0}{shape parameter for the Beta(a0, b0) mixing prior.}

\item{tau_a0}{shape parameter for the inverse-Gamma(a0, b0) prior on the variance, tau^2.}

\item{tau_b0}{scale parameter for the inverse-Gamma(a0, b0) prior on the variance, tau^2.}

\item{mu}{initial values of mu, the means of the variational family.}

\item{s}{initial values of s, the std. dev of the variational family.}

\item{g}{initial values of g, the group inclusion probabilities of the variational family.}

\item{track_elbo}{track the evidence lower bound (ELBO).}

\item{track_elbo_every}{the number of iterations between computing the ELBO.}

\item{track_elbo_mcn}{number of Monte-Carlo samples to compute the ELBO.}

\item{niter}{maximum number of iteration to run the algorithm for.}

\item{tol}{convergence tolerance.}

\item{verbose}{print additional information.}
}
\value{
The program output is a list containing:
\item{mu}{the means for the variational posterior.}
\item{s}{the std. dev. or covaraince matrices for the variational posterior.}
\item{g}{the group inclusion probabilities.}
\item{beta_hat}{the variational posterior mean.}
\item{tau_hat}{the mean of the variance term.}
\item{tau_a}{the shape parameter for the variational posterior of tau^2. This is an inverse-Gamma(tau_a0, tau_b0) distribtuion.}
\item{tau_b}{the scale parameter for the variational posterior of tau^2.}
\item{parameters}{a list containing the model hyperparameters.}
\item{converged}{a boolean indicating if the algorithm has converged.}
\item{iter}{the number of iterations the algorithm was ran for.}
}
\description{
Fit high-dimensional group-sparse regression models
}
\section{Details}{
 TODO
}

\examples{
n <- 100
p <- 1000
gsize <- 5
groups <- c(rep(1:(p/gsize), each=gsize))

X <- matrix(rnorm(n * p), nrow=n, ncol=p)
b <- c(rep(0, gsize), rep(-4, gsize), rep(8, gsize), rep(0, p - 3 * gsize))
y <- X \%*\% b + rnorm(n, 0, 1)

f <- gsvb.fit(y, X, groups)

plot(f$beta_hat, col=4, ylab=expression(hat(beta)))
points(b, pch=20)


}
