% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{gsvb.predict}
\alias{gsvb.predict}
\title{Sample from the posterior predictive distribution}
\usage{
gsvb.predict(
  fit,
  newdata,
  samples = 10000,
  quantiles = c(0.025, 0.975),
  return_samples = FALSE
)
}
\arguments{
\item{fit}{the fit model.}

\item{newdata}{input feature matrix.}

\item{samples}{number of Monte-Carlo samples.}

\item{quantiles}{quantiles to return}

\item{return_samples}{return all the samples}
}
\value{
a list with the mean of the posterior predictive, the quantiles, and optionally the samples.
}
\description{
Sample from the posterior predictive distribution
}
\examples{
n <- 100
p <- 1000
gsize <- 5
groups <- c(rep(1:(p/gsize), each=gsize))

X <- matrix(rnorm(n * p), nrow=n, ncol=p)
b <- c(rep(0, gsize), rep(-4, gsize), rep(8, gsize), rep(0, p - 3 * gsize))
y <- X \%*\% b + rnorm(n, 0, 1)

f <- gsvb.fit(y, X, groups)
gsvb.predict(f, groups, X) 

}
