% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.r
\name{gsvb.sample}
\alias{gsvb.sample}
\title{Sample from the variational posterior distribution}
\usage{
gsvb.sample(fit, samples = 10000)
}
\arguments{
\item{fit}{the fit model.}

\item{samples}{number of samples}
}
\value{
a list containing:
\item{beta}{a matrix of samples from beta}
\item{tau}{vector of samples from tau (only for linear model)}
}
\description{
Sample from the variational posterior distribution
}
\examples{
n <- 100
p <- 1000
gsize <- 5
groups <- c(rep(1:(p/gsize), each=gsize))

X <- matrix(rnorm(n * p), nrow=n, ncol=p)
b <- c(rep(0, gsize), rep(-4, gsize), rep(8, gsize), rep(0, p - 3 * gsize))
y <- X \%*\% b + rnorm(n, 0, 1)

f <- gsvb.fit(y, X, groups)
s <- gsvb.sample(f)
hist(s$beta[10, ])

}
